Screen control register Vertical raster scroll incorrect
I noticed you had $38 which causes the vic scroll to be "0" which pushes the image up on the screen. The value should be $3B (setting the lower three bits to %011), which is the default "neutral" scroll for the system. The Hires FLI demo also uses this $38 value but that scroll is changed the next frame so it doesn't cause any issues.