Added C64 Multi FLI (MCI) support
The C64 Multi FLI uses a 4x1 screen block size allowing a background
color, two screen ram color choices, and a third "color block" color
choice option. The color block is 8x8 in size and stored in the
VIC color ram in a fixed memory location. The C64 screen
ram color choices can be adjusted per scan line (as done in FLI) but
the color block ram address cannot be dynamically changed). The
sample code changes screen ram address banks per Y scan line
but leaves the color block ram alone as insufficient remaining CPU
time exists to do any meaningful changes to the color block ram.

Both C64 Multi FLI and Hires FLI have the "bad scan line" bug
present in the example source code. The first 3 character columns
on the PAL screen will not have color information because the VIC is
stalled on CPU wait cycles and thus the character color data is
not available when the scan line starts to draw (and the VIC uses
pixel colors that happen to remain in the registers at
the time of the bad scan line is invoked). This is a known bug
with the FLI technique without a solution. Most developers work
around thus limitation by filling the screen character ram
with a 00 pixel pattern so the pixels are transparent and
display the standard background color instead. Emulators
(and SCPU) however can "fix" this bug and display the full
160x200 picture but this fix is outside the scope of tool
and example code.