Added code to handle VIC FLI code properly
Related to:
https://github.com/sehugg/dithertron/issues/10

In FLI mode the raster cycle accuracy has to be exact to the
raster beam position aligned with the clock cycle. This is a
complex procedure using a double IRQ.

Even when done correctly, FLI isn't perfect as the left side of the
screen lacks proper coloring for 24 pixels.This is known as
the 'FLI bug" and it's present in both multi-color mode and
hires mode. Without accurate clock cycles that fli bug error
on the left side increases (until nothing gets displayed at all).

In FLI mode, the only color choices available for the left side of
the screen for multi-color mode is: background, light grey or
orange (yuck). In hires mode, it's light grey and that's it.

The solution is to either ignore the bug and use a patched
emulator which doesn't have the FLI bug (even though real
C64 hardware has the bug), work with those awful colors and
accept a glitched left side, or blank the left (and optionally the
left and right). That's why so many choices exist for FLI modes
since what is the right option is picture/usage dependent.

Other changes made:
- all the C64 modes uses the VIC canvas since the code not
only capable of each mode, but it has the best logic for handling
the FLI bug
- replaced the FLI assembly code with cycle accurate code which
is compatible with both PAL and NTSC machines
- clean up the settings  bit as the color block and normal block aren't
that different except their dimensions
- fixed the background color choices for normal non FLI modes
- added all the FLI options in the pull down list... if someone wants to
instead make these sub-choices show up only in FLI mode, be my guest!