Vic 20 dithering support added (preliminary)
The Vic 20 has some unique attributes that are similar but different
than the Commodore 64. The Vic 20 doesn't have a true graphics
mode but instead relies on character re-imaging and color choices
to make a pseudo bitmap. Combined with the Vic 20's flexible
screen sizes, this makes an interesting target.

Some key differences added:
- In multi mode, the Vic 20 has the choice of 4 color options
just like the C64, except three of the choices are the background,
border, and auxiliary color. These are global colors, and then
each block gets the choice of its own foreground color.
- In hires mode, the choice is the background or a foreground
color for the block
- In multi-mode the palette is fully available for the background,
and aux color but a palette subset is used for the aux color and
the foreground color.
- In hires mode, the palette is restricted for the background and
the foreground color.

Code changes (some refactoring):
- The global color selection is improved, because the choice
of global colors is so critical for the vic, the background, aux,
and border colors are selected by pixels, whereas commodore 64
only uses the background (at times), the logic now first uses
histogram usage to pick the top N colors based on usage if the
color is useable by the image's pixels. The remaining colors
that are needed to have compatible colors are based on the
intensity and are selected for palette compatibility to the
image rather than overall usage.
- the palette options got a lot more complex in the settings
so those got split out into their own dithering settings with
appropriate defaults that would mean the rest of the settings
don't need to specify anything and they will still work
- the pixel color choices now respect the palette subset that
can happen in the case of the Vic 20
- added VIC 20 palette colors (there are some differences between
systems but this is true of the C64 as well)
- added export bin routines for the Vic 20 which are likely to
produce nothing much of useful value given they aren't tested
against working code
- misc typos and a few inconsistencies in the code (code could use
a good auto-format tool as I find the editors liking to do it
the way they think the code should be not how the code is)