Added dark and bright modes to Zx Spectrum
The Zx spectrum has interesting palette restrictions. Each pixel
can choose from the bright part of the palette or the dark part
of the palette in a cell, but no cell can choose both bright and
dark at the same time.

The previous code selected from palette that was bright only
and then forced a bright mode. The new logic understands the
possible dark vs light choice and then chooses the appropriate
dark/light choice per cell.

New options also exist to force a light or dark mode if the
user has a preference for one palette or the other (for whatever
reasons they might have).

Changes to the code:
- the spectrum palette was expanded to include dark and bright
- systems were added to allow choices between the dark and bright
- the cell exporter was made more generic to allow x/y mapping
to param/cell to be generic enough to handle the oddity that is
the Zx Spectrum memory layout
- the zs canvas logic was made to understand the dark vs bright
selection choice