Fix: TMS9918 coleco system transparency
The Coleco TMS9918 has a special transparency pixel in the palette
which in the dithertron was specified as RGB of black. The trouble is
that when black is mapped in the export, it becomes the background
transparency color. This fixes the issue by remapping the transparent
pixel to black.

Changes to the code:
- the TMS9918 was using the older character exporter routine, which
is now using the common character exporter routine
- made the extra params a setting of the system (for more generic
usage in routines)
- removed older exportCharMemory routine
- as the TMS9918 does not map the color ram in a linear fashion
like the other systems, added a color ram/color block ram mapper
to make it easier to re-map parameters onto the color space
while still using the common export routines
- remap transparent onto black color in export