Added Intellivision STIC dithering
The intellivision has two interesting modes. The foreground and
background mode allows the choice of two colors for an 8x8 area
but the downside is that the characters are restricted to gram, of
which there are only 64 of 8x8 pixels. This makes a rather small
image.

The modes use a color stack, which allows for up to 256 characters
(but only 240 are needed to fill the screen) and requires
GROM image replacement by overriding the ROM bus. The
foreground color is restricted but registers are used to
chose the current background color in a rotating wheel of
colors. The choice of the color is not selectable, as each character
has the choice of picking the current color in the stack, or
choosing the next color in the stack. Thus the value of the
color stack is diminished.

The color stack code figures out the most useful colors to
use for the color stack (and the most likely to be chosen),
and then figures out the best order combination to use for
the color stack. An interesting extension might be to force
only a single color to be used, like a global background
color (TBD if that's helpful).

Changes to code:
- as this new code heavily relies upon the same mechanics
as the Vic canvas, a new block canvas base was created
- the new base class makes use of the settings to automatically
adjust to most common block color scenarios without changes
- the overrides exist to help facilitate strange approaches (such
as the STIC), while still keeping a huge common base of the code
- the export routines had an overhaul so they could handle the
variety of strangeness that can happen (including the STIC
which has to export GROM characters, and then duplicate
GROM characters into GRAM to get extended color options)
- an extended set of configured and defaulted settings is used
for the block canvas to avoid a lot of recalculations in the export
- bunch of helper routines to normalize the encoding of data
into params and extract out of params
- split out params as a single array into a color params (i.e.
what is known as screen params in commodores), color block
params, cell params, and extra params
- instead of adding more and more message values into the export,
a single "content" is exported allowing a bulk transfer of much more
useful values to the export routines
- fixed the commit bug where the update and commit get out of sync