Added dithering of SNES
This adds 32x32 grid of 2bpp to 8bpp as well as direct color modes for
the SNES. This does not allow exporting of the data but it's a first
step towards that goal. SNES is considered experimental and untested
if the result is usable in the real world (although it is coded to be
proper).

Changes to the code:
- added SNES modes
- added SNES direct color mode
- refactored the export code to be more generic
- fixed sync bug between canvas and indexed
- added full pallette mode to block canvas
- fixed vic20 fill choice (to not default to background)
- fixed vic20 to not sort the colors of two choices, but to instead
pick the top color only
- indexed needs to support > 256 values (SNES has 11 bit colors)
- made range routine faster by pre-allocating array