Updated console.assert to assert outside lock
On the unit tests, the console.assert routine is extremely slow
because it goes through some kind of slow global lock.

As the console assertion should only happen in a failure condition,
there's no reason to obtain a lock to perform the test, so the change
only calls the assert during the failure should work well.